modification in route generation

+++ src/jmetal/operators/mutation/RouteSetTELMutation.java
@@ -71,7 +71,7 @@ public class RouteSetTELMutation extends Mutation
         int minNode = prob.ins.getMinNode();
         int maxNode = prob.ins.getMaxNode();
         int Vertices = prob.ins.getNumOfVertices();
-        int[][] time = prob.getTime();
+        double[][] time = prob.getTime();
         int I = numberOfNodes(maxNode, rs.size());
         Grph reducedGraph = new InMemoryGrph();
         // reducedGraph.display();
